+++ src/main/java/com/ham/netnovel/novel/service/NovelService.java
@@ -2,14 +2,15 @@ package com.ham.netnovel.novel.service;
 
 import com.ham.netnovel.common.exception.ServiceMethodException;
 import com.ham.netnovel.novel.Novel;
-import com.ham.netnovel.novel.data.NovelSearchType;
 import com.ham.netnovel.novel.dto.*;
 import org.springframework.data.domain.Pageable;
+import org.springframework.security.access.AccessDeniedException;
 import org.springframework.web.multipart.MultipartFile;
 
 import java.time.LocalDateTime;
 import java.util.List;
 import java.util.Map;
+import java.util.NoSuchElementException;
 import java.util.Optional;
 
 public interface NovelService {
@@ -45,61 +46,43 @@ public interface NovelService {
 
 
 
+
     /**
-     * DB에 저장된 Novel 삭제.
-     * @param novelDeleteDto 삭제 정보가 담긴 dto
+     * 주어진 소설 ID를 사용하여 소설의 상태를 삭제상태로 변경하는 메서드 입니다.
+     *
+     *
+     * <p>
+     * 소설 ID 로 소설을 조회합니다.
+     * 소설 작가와 삭제 요청자의 providerId 가 동일한경우 소설을 DELETED_BY_USER 로 변경하고 DB에 저장합니다.
+     * </p>
+     * @param novelDeleteDto 소설 삭제에 필요한 정보가 담긴 DTO
+     * @throws NoSuchElementException 해당 ID의 소설이 존재하지 않을 경우 발생
+     * @throws AccessDeniedException 접근 권한이 없을 경우 발생
+     * @throws ServiceMethodException 권한 검증 도중 예외가 발생했을 경우 발생
      */
     void deleteNovel(NovelDeleteDto novelDeleteDto);
 
     /**
-     * Novel 엔티티를 대략적으로 설명하는 데이터를 가져오는 메서드
-     * @param novelId Novel PK 값
-     * @return NovelInfoDto
+     * 주어진 소설 ID를 사용하여 소설 정보를 조회합니다.
+     * <p>
+     * 이 메서드는 제공된 소설 ID를 통해 DB에서 소설 데이터를 조회합니다.
+     * 만약 DB에 소설 정보가 없으면{@link NoSuchElementException} 로 예외를 던집니다.
+     *</p>
+     *
+     * @param novelId 조회할 소설의 ID
+     * @return 조회된 소설의 정보가 담긴 {@link NovelInfoDto}
+     * @throws NoSuchElementException 소설 정보를 찾을 수 없을 때 발생
+     * @throws ServiceMethodException 메서드 실행 중 변환 과정에서 예외가 발생할 때 발생
      */
     NovelInfoDto getNovelInfo(Long novelId);
 
-    //ToDo 메서드 기능 통합 후 삭제
-    /**
-     * DB의 저장된 모든 Novel 리스트를 최신순으로 가져오는 메서드.
-     * 페이지 단위로 일부 리스트만 가져온다.
-     * @param pageable page number, page size
-     * @return
-     */
-    List<NovelInfoDto> getNovelsRecent(Pageable pageable);
 
-    /**
-     * 유저의 선호작 Novel 리스트 반환.
-     * @param providerId 유저 PK 값.
-     * @return List<Novel>
-     */
-    List<Novel> getFavoriteNovels(String providerId);
 
     /**
      * 별점 점수가 있는 Novel의 id값들을 List로 반환하는 메서드
      * @return List Long 타입으로 반환
      */
     List<Long> getRatedNovelIds();
-
-
-    /**
-     * 주어진 기간에 따라 소설의 랭킹을 조회하여 해당 페이지의 소설 정보를 반환합니다.
-     *
-     * <p>이 메서드는 다음과 같은 작업을 수행합니다:</p>
-     * <ul>
-     *     <li>페이지 번호와 페이지 크기를 사용하여 데이터의 시작 인덱스와 끝 인덱스를 계산합니다.</li>
-     *     <li>Redis에서 주어진 기간에 해당하는 소설 랭킹 데이터를 가져옵니다.</li>
-     *     <li>가져온 데이터에서 소설 ID를 추출합니다.</li>
-     *     <li>추출한 소설 ID를 사용하여 소설 엔티티를 조회하고, 랭킹 순서로 정렬하여 DTO로 변환합니다.</li>
-     * </ul>
-     *
-     * <p>작업 중 예외가 발생하면 {@link ServiceMethodException}이 발생합니다.</p>
-     *
-     * @param period 소설 랭킹을 조회할 기간. "daily", "weekly", "monthly" 중 하나로 지정합니다.
-     * @param pageable 페이지 정보. 페이지 번호와 페이지 크기를 포함합니다.
-     * @return 주어진 페이지와 기간에 해당하는 소설 정보를 담은 {@link List}입니다.
-     * @throws ServiceMethodException 메서드 실행 중 오류가 발생한 경우 발생합니다.
-     */
-    List<NovelListDto> getNovelsByRanking(String period, Pageable pageable);
 
 
     /**
@@ -167,43 +150,8 @@ public interface NovelService {
     Map<Long, LocalDateTime> getNovelWithLatestEpisodeCreateTime(Pageable pageable);
 
 
-    /**
-     * 주어진 정렬 기준과 페이지 정보를 바탕으로 소설 목록을 조회하는 메서드입니다.
-     *
-     * <p>
-     * 이 메서드는 기본적으로 소설 목록을 조회수 기준으로 정렬합니다.
-     * </p>
-     *
-     * <p>
-     * 사용자가 제공한 정렬 기준에 따라, 소설 목록을 좋아요 순 또는 최신순으로 정렬할 수 있습니다.
-     * 조회된 소설 목록의 썸네일 URL은 S3 서비스의 CloudFront URL로 변환됩니다.
-     * </p>
-     * @param sortOrder 소설 목록의 정렬 기준을 나타내는 문자열입니다.
-     *                  "favorites"는 좋아요 순, "latest"는 최신순, 기본값은 "view"로 조회수 순입니다.
-     * @param pageable 페이지 정보 및 페이징 조건을 포함하는 {@link Pageable} 객체입니다.
-     *
-     * @return List<NovelListDto> 소설 목록을 포함한 DTO 리스트를 반환합니다.
-     *         각 소설의 썸네일 URL은 CloudFront URL로 변환되어 반환됩니다.
-     * @throws ServiceMethodException 검색 중 예외 발생 시 예외를 던집니다.
-     */
-    List<NovelListDto> getNovelsBySearchCondition(String sortOrder, Pageable pageable, List<Long> tagIds);
 
-    /**
-     * 검색어를 기반으로 소설 목록을 검색하여 반환합니다.
-     * <p>
-     * 검색된 소설 정보를 NovelListDto로 변환하여 페이징 처리된 결과를 반환합니다.
-     * 조회된 소설 목록의 썸네일 URL은 S3 서비스의 CloudFront URL로 변환됩니다.
-     * </p>
-     *
-     * @param searchWord 유저가 입력한  검색어 {@link String} 객체
-     * @param pageable 페이지 정보 및 페이징 조건을 포함하는 {@link Pageable} 객체입니다.
-     * @return List<NovelListDto> 소설 목록을 포함한 DTO 리스트를 반환합니다.
-     *         각 소설의 썸네일 URL은 CloudFront URL로 변환되어 반환됩니다.     *
 
-     * @throws ServiceMethodException 검색 중 예외 발생 시 예외를 던집니다.
-     */
-    List<NovelListDto> getNovelsBySearchWord(String searchWord,
-                                             NovelSearchType novelSearchType,
-                                             Pageable pageable);
+
 
 }
